fix: 베팅 분석 결과가 0인 문제 해결

+++ aiku/src/main/java/konkuk/aiku/service/GroupService.java
@@ -193,7 +193,7 @@ public class GroupService {
                     .filter(betting -> betting.getResultType().equals(ResultType.LOSE))
                     .count();
 
-            int winningRate = totalBettingCount == 0 ? 0 : (winCount / totalBettingCount) * 100; // 백분율
+            int winningRate = totalBettingCount == 0 ? 0 : (int) (((double) winCount / totalBettingCount) * 100); // 백분율
 
             analyticsResult.add(
                     AnalyticsBettingServiceDto.createDto(userInGroup, winningRate)
